account-data backend updates and corrections
