+++ backend/vizzy-backend/src/user/user.controller.ts
@@ -1,6 +1,7 @@
-import { Controller, Get, Param } from '@nestjs/common';
+import { Controller, Get, Param, UseGuards } from '@nestjs/common';
 import { UserService } from './user.service';
 import { User } from './models/user.model';
+import { JwtAuthGuard } from '../auth/guards/jwt.auth.guard';
 
 @Controller('user')
 export class UserController {
@@ -12,7 +13,7 @@ export class UserController {
   }
 
   @Get('me')
-  //@UseGuards(JwtAuthGuard)
+  @UseGuards(JwtAuthGuard)
   async getMe(@Param('id') id: string): Promise<User | null> {
     return this.userService.getUserById(id);
   }
